# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/mangas": {
            "get": {
                "description": "Retrieve all mangas with sorting and pagination options.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mangas"
                ],
                "summary": "Get All Mangas Sorted and Paginated",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of results per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.MangaDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new manga in the system.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mangas"
                ],
                "summary": "Create Manga",
                "parameters": [
                    {
                        "description": "Manga Insert DTO",
                        "name": "mangaInsertDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.MangaInsertDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Manga Created",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Data",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    }
                }
            }
        },
        "/api/mangas/author/{author}": {
            "get": {
                "description": "Retrieve all mangas written by a specific author.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mangas"
                ],
                "summary": "Get Mangas by Author",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Author Name",
                        "name": "author",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.MangaDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Author Name",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "No mangas found for this author",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    }
                }
            }
        },
        "/api/mangas/demography/{demography}": {
            "get": {
                "description": "Retrieve all mangas under a specific demography.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mangas"
                ],
                "summary": "Get Mangas by Demography",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Demography Type",
                        "name": "demography",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of results per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.MangaDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Demography",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "No mangas found for this demography",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    }
                }
            }
        },
        "/api/mangas/genre/{genre}": {
            "get": {
                "description": "Retrieve all mangas under a specific genre.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mangas"
                ],
                "summary": "Get Mangas by Genre",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Genre",
                        "name": "genre",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of results per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.MangaDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Genre",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "No mangas found for this genre",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    }
                }
            }
        },
        "/api/mangas/name/{name}": {
            "get": {
                "description": "Retrieve all mangas that match a specific name pattern.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mangas"
                ],
                "summary": "Get Mangas by Name Pattern",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Manga Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of results per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.MangaDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Name",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "No mangas found for this name",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    }
                }
            }
        },
        "/api/mangas/{mangaId}": {
            "get": {
                "description": "Retrieve a manga by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mangas"
                ],
                "summary": "Get Manga by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Manga ID",
                        "name": "mangaId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/dtos.MangaDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid Manga ID",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Manga not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing manga by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mangas"
                ],
                "summary": "Update Manga",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Manga ID",
                        "name": "mangaId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Manga Insert DTO",
                        "name": "mangaInsertDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.MangaInsertDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Manga Updated",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Manga ID",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Manga not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a manga by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mangas"
                ],
                "summary": "Delete Manga",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Manga ID",
                        "name": "mangaId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Manga Deleted",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Manga ID",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Manga not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.MangaDTO": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "author": {
                    "type": "string"
                },
                "chapters": {
                    "type": "integer"
                },
                "cover_image_url": {
                    "type": "string"
                },
                "demograhpy": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "publication_date": {
                    "type": "string"
                },
                "publisher": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "volume": {
                    "type": "integer"
                }
            }
        },
        "dtos.MangaInsertDTO": {
            "type": "object",
            "required": [
                "author",
                "chapters",
                "demograhpy",
                "genres",
                "title",
                "volume"
            ],
            "properties": {
                "author": {
                    "type": "string"
                },
                "chapters": {
                    "type": "integer"
                },
                "cover_image_url": {
                    "type": "string"
                },
                "demograhpy": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "publication_date": {
                    "type": "string"
                },
                "publisher": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "volume": {
                    "type": "integer"
                }
            }
        },
...
        "utils.ApiResponse": {
            "description": "Represents a standardized API response structure.",
            "type": "object",
            "properties": {
                "code": {
                    "description": "The HTTP status code",
                    "type": "integer"
                },
                "data": {
                    "description": "The data returned by the API"
                },
                "msg": {
                    "description": "A message providing additional context",
                    "type": "string"
                },
                "success": {
                    "description": "Indicates if the request was successful",
                    "type": "boolean"
                },
                "timestamp": {
                    "description": "The time at which the response was generated",
                    "type": "string"
                }
            }
        }
    }
}